[C-08]: Move Component of this container to components folder
