+++ src/components/CenterDiv/CenterDiv.js
@@ -1,6 +1,8 @@
 import styled from "styled-components";
 
-export const CenterDiv = styled.div`
+const CenterDiv = styled.div`
   display: flex;
   justify-content: center;
 `;
+
+export default CenterDiv;
