+++ src/containers/PracticeAnswer/components/Shift.js
@@ -1,0 +1,20 @@
+import styled from "styled-components";
+
+// Components
+import CenterDiv from "../../../components/CenterDiv/CenterDiv";
+
+export const ShiftDiv = styled(CenterDiv)`
+  min-height: 72px;
+`;
+
+export const ShiftLeft = styled.img`
+  alt: "Correct Backward";
+  height: 40px;
+  margin-right: 32px;
+`;
+
+export const ShiftRight = styled.img`
+  alt: "Correct Backward";
+  height: 40px;
+  margin-left: 32px;
+`;
