+++ src/containers/PracticeAnswer/components/Sign.js
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+const Sign = styled.img`
+  alt: "Correct Sign";
+  height: 72px;
+  margin-top: 68px;
+`;
+
+export default Sign;
